refactor(step2): 디렉토리 변경
